// src/core/enum/currentUser.enum.ts
export enum CurrentUserEnum {
    CURRENT_USER = 'currentUser'
}

export enum GetWalkValues {
    GREEN = 'green',
    RED = 'red',
    HIGHSCORE = 'highscore',
    SCORE = 'score',
    LEFT = 'left',
    RIGHT = 'right'
}

export enum HomeLogin {
    ENTER_CREATE_NAME = 'Enter / Create player',
    NAME = 'Name',
    ENTER = 'Enter'
}